Change the Coyote class requiring the user manual to always be there.
The Coyote now raises an error in case there's no user manual for the product from the API.

With this change, even though the manual is still coming in the user_guide attribute and the app is still not aware of it,
there's no false assumption anymore since an error will be raised by the Coyote when it sees that the user manual is missing.
During runtime we would see this error being popped up and would be able to track down the cause of the missing user manual.

Since the Coyote is the direct consumer and requester for the product details, would be easy to track this down.
The next commit simulates a more difficult scenario where more layers of indirections are in place.