Check request expectations when performing requests.
Add simple way to check whether the JSON response contains the required attributes
for the app. An error is raised as soon as any of the expectations is not met,
(in this case, the missing user_manual) showing a nice error message so that we know the
exact root cause of the issue.